Screensaver[BETA]
The "Main" method. Call "startidlescreen" (No initialization required since it is STATIC) in the master main method after there is no action for a certain time(the other group developing the master program has to implement this requirement). This method will stop when the mouse is clicked, so don't worry about attaching another timer since it is already implemented inside.