Still implementing zoombuttons
There has been multiple issues with whether to use jlabel or jpanels, and images not showing up when repainting. I am going to go with the jlabel route as it is easier, but jpanel would be much more flexible if other groups have multiple components on screen. This code currently has a resizing and repainting crash error when pressing the + button.