Fixed rare zoom error
When user zooms in then out, there may still be a few pixels difference between the zoomedimage and the original image, allowing zooming out to happen again, resulting in the code trying to create an Image larger than possible. I added a try/catch to set the zoomedimage as the maximum size if this error is encountered.